cli: move the demo cluster's tests to their own file
Release note: None
